removed fix from onehot

+++ src/super_gradients/training/utils/segmentation_utils.py
@@ -47,7 +47,7 @@ def to_one_hot(target: torch.Tensor, num_classes: int, ignore_index: int = None)
     """
     num_classes = num_classes if ignore_index is None else num_classes + 1
 
-    one_hot = F.one_hot(target.type(torch.long), num_classes).permute((0, 3, 1, 2))
+    one_hot = F.one_hot(target, num_classes).permute((0, 3, 1, 2))
 
     if ignore_index is not None:
         # remove ignore_index channel
